+++ method/stLearn/config/config_11.json
@@ -1,0 +1,1 @@
+{"image": true, "normalize": false, "method": "louvain", "n_comps": 15, "n_neighbors": 15, "resolution": 1}
