+++ method/stLearn/config/config_12.json
@@ -1,0 +1,1 @@
+{"image": false, "method": "kmeans", "n_comps": 15}
