+++ method/stLearn/config/config_13.json
@@ -1,0 +1,1 @@
+{"image": false, "method": "louvain", "n_comps": 15, "n_neighbors": 15, "resolution": 1}
