+++ method/stLearn/config/config_2.json
@@ -1,1 +1,1 @@
-{"method": "kmeans", "n_comps": 15}
+{"image": true, "normalize": true, "weights": "physical_distance", "method": "kmeans", "n_comps": 50}
