+++ method/stLearn/config/config_3.json
@@ -1,1 +1,1 @@
-{"method": "kmeans", "n_comps": 25}
+{"image": true, "normalize": true, "weights": "physical_distance", "method": "louvain", "n_comps": 15, "n_neighbors": 15, "resolution": 1}
