+++ method/stLearn/config/config_6.json
@@ -1,1 +1,1 @@
-{"method": "louvain", "n_comps": 15, "n_neighbors": 10, "res": 0.5}
+{"image": true, "normalize": true, "weights": "physical_distance", "method": "louvain", "n_comps": 15, "n_neighbors": 15, "resolution": 1.2}
