+++ method/stLearn/config/config_7.json
@@ -1,1 +1,1 @@
-{"method": "louvain", "n_comps": 15, "n_neighbors": 10, "res": 0.7}
+{"image": true, "normalize": true, "weights": "weights_matrix_pd_md", "method": "kmeans", "n_comps": 15}
