+++ method/stLearn/config/config_9.json
@@ -1,1 +1,1 @@
-{"method": "louvain", "n_comps": 15, "n_neighbors": 10, "res": 1.2}
+{"image": true, "normalize": true, "weights": "weights_matrix_all", "method": "louvain", "n_comps": 15, "n_neighbors": 15, "resolution": 1}
