+++ method/stLearn/stLearn.yml
@@ -1,16 +1,15 @@
 channels:
   - conda-forge
 dependencies:
-  - python=3.8
-  - bokeh>= 2.4.2
-  - click>=8.0.4
-  - leidenalg
-  - louvain
-  - numpy>=1.18,<1.22
-  - Pillow>=9.0.1
-  - scanpy>=1.8.2
-  - scikit-image>=0.19.2
-  - tensorflow
-  - pip
-  - pip:
-    - stLearn==0.4.12
+  - python=3.10.13
+  - bokeh=3.3.3
+  - click=8.1.7
+  - leidenalg=0.10.2
+  - louvain=0.8.1
+  - numba=0.57.1
+  - numpy=1.21.6
+  - pillow=10.0.0
+  - scanpy=1.9.7
+  - scikit-image=0.20.0
+  - tensorflow=2.11.1
+  - stLearn=0.4.12
